requirements.txt
flask >= 2.1.1
python-telegram-bot >= 13.11
google-cloud-datastore >= 2.5.1